+++ src/Application.hpp
@@ -161,7 +161,7 @@ private:
     FfzBadges *const ffzBadges{};
     SeventvBadges *const seventvBadges{};
     UserDataController *const userData{};
-    ISoundController *const sound{};
+    std::unique_ptr<ISoundController> sound;
     TwitchLiveController *const twitchLiveController{};
     std::unique_ptr<PubSub> twitchPubSub;
     std::unique_ptr<TwitchBadges> twitchBadges;